requirements.txt
alembic==1.12.1
blinker==1.6.3
certifi==2025.4.26
click==8.1.8
colorama==0.4.6
dnspython==2.3.0
dominate==2.9.1
elastic-transport==8.13.1
elasticsearch==8.14.0
email-validator==1.3.1
Flask==2.2.5
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.1.0
Flask-Moment==1.0.6
flask-sqlalchemy==3.0.5
Flask-WTF==1.1.1
greenlet==3.1.1
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
jinja2==3.1.6
Mako==1.2.4
MarkupSafe==2.1.5
packaging==24.0
PyJWT==2.8.0
python-dotenv==0.21.1
SQLAlchemy==2.0.41
typing-extensions==4.7.1
urllib3==2.0.7
visitor==0.1.3
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
